Add files via upload
The side-by-side box plot indicate that the blood pressure is a big indicator of whether the patients should be admitted immediately. In cases where the blood pressure is high the medical doctor's decision is generally uniform.
The histogram results indicate that many patients have a blood pressure of between 50 and 150.